requirements.txt
blis==0.4.1
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
cymem==2.0.2
en-core-web-md==2.1.0
idna==2.8
joblib==0.14.0
kiwisolver==1.1.0
llvmlite==0.29.0
matplotlib==3.1.1
murmurhash==1.0.2
nltk==3.4.5
numba==0.45.1
numpy==1.17.2
plac==0.9.6
preshed==3.0.2
psutil==5.6.3
pyparsing==2.4.2
python-dateutil==2.8.0
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
spacy==2.1.8
# Note that in this project en-core-web-md model is used
srsly==0.1.0
thinc==7.1.1
tqdm==4.36.1
umap-learn==0.3.10
urllib3==1.25.6
wasabi==0.2.2
